+++ core/uikit/src/commonMain/kotlin/ru/starfactory/core/uikit/view/FloatingActionButtons.kt
@@ -1,0 +1,61 @@
+package ru.starfactory.core.uikit.view
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.border
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.shape.CornerSize
+import androidx.compose.material.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun POutlinedFloatingActionButton(
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+    interactionSource: MutableInteractionSource = remember { MutableInteractionSource() },
+    shape: Shape = MaterialTheme.shapes.small.copy(CornerSize(percent = 50)),
+    backgroundColor: Color = Color.Transparent,
+    contentColor: Color = contentColorFor(backgroundColor),
+    elevation: FloatingActionButtonElevation = FloatingActionButtonDefaults.elevation(0.dp, 0.dp, 0.dp, 0.dp),
+    borderStroke: BorderStroke = BorderStroke(2.dp, MaterialTheme.colors.surface),
+    content: @Composable () -> Unit
+) {
+    PTextFloatingActionButton(
+        onClick,
+        modifier
+            .border(borderStroke, shape),
+        interactionSource,
+        shape,
+        backgroundColor,
+        contentColor,
+        elevation,
+        content
+    )
+}
+
+@Composable
+fun PTextFloatingActionButton(
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+    interactionSource: MutableInteractionSource = remember { MutableInteractionSource() },
+    shape: Shape = MaterialTheme.shapes.small.copy(CornerSize(percent = 50)),
+    backgroundColor: Color = Color.Transparent,
+    contentColor: Color = contentColorFor(backgroundColor),
+    elevation: FloatingActionButtonElevation = FloatingActionButtonDefaults.elevation(0.dp, 0.dp, 0.dp, 0.dp),
+    content: @Composable () -> Unit
+) {
+    FloatingActionButton(
+        onClick,
+        modifier,
+        interactionSource,
+        shape,
+        backgroundColor,
+        contentColor,
+        elevation,
+        content
+    )
+}
